make sendDonorThankYouEmail sg helper
